~ Rename package to bittorrent.
It seems like 'network-bittorrent' is too long and annoying: everyone
knows that bittorrent is a network protocol and there is no purpose to
prefix it with 'network'. Moreover hackage do not have package with
the our new name.

So I think just 'bittorrent' is a more suitable name for the package.